[Design] 일지 작성 템플릿에 태그 입력 창 추가
