+++ app/src/main/java/com/kittyandpuppy/withallmyanimal/mypage/MypageFragment.kt
@@ -1,0 +1,36 @@
+package com.kittyandpuppy.withallmyanimal
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.kittyandpuppy.withallmyanimal.R
+
+class MypageFragment : Fragment() {
+    private var param1: String? = null
+    private var param2: String? = null
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        arguments?.let {
+        }
+    }
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_mypage, container, false)
+    }
+
+    companion object {
+        @JvmStatic
+        fun newInstance(param1: String, param2: String) =
+            MypageFragment().apply {
+                arguments = Bundle().apply {
+                }
+            }
+    }
+}
