+++ app/src/main/res/layout/activity_mypage_behavior.xml
@@ -173,17 +173,47 @@
                 app:layout_constraintTop_toBottomOf="@id/tv_mypage_behavior_title" />
 
             <TextView
+                android:id="@+id/tv_mypage_behavior_tag"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:text="태그(최대 3줄)"
+                android:fontFamily="@font/cafe24"
+                android:textColor="@color/black"
+                android:textSize="14sp"
+                android:layout_marginStart="20dp"
+                android:layout_marginTop="20dp"
+                app:layout_constraintTop_toBottomOf="@id/etv_mypage_behavior"
+                app:layout_constraintStart_toStartOf="@id/view_mypage_behavior"
+                />
+
+            <EditText
+                android:id="@+id/etv_mypage_behavior_tag"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:layout_marginHorizontal="50dp"
+                android:layout_marginTop="10dp"
+                android:background="@drawable/mypage_diary_edittext"
+                android:fontFamily="@font/cafe24"
+                android:hint="태그를 작성해주세요"
+                android:padding="20dp"
+                android:textColor="@color/grey"
+                android:textSize="13sp"
+                app:layout_constraintEnd_toEndOf="@id/view_mypage_behavior"
+                app:layout_constraintStart_toStartOf="@id/view_mypage_behavior"
+                app:layout_constraintTop_toBottomOf="@id/tv_mypage_behavior_tag" />
+
+            <TextView
                 android:id="@+id/tv_mypage_behavior_review"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:layout_marginStart="20dp"
-                android:layout_marginTop="30dp"
+                android:layout_marginTop="20dp"
                 android:fontFamily="@font/cafe24"
                 android:text="후기"
                 android:textColor="@color/black"
                 android:textSize="14sp"
                 app:layout_constraintStart_toStartOf="@id/view_mypage_behavior"
-                app:layout_constraintTop_toBottomOf="@id/etv_mypage_behavior" />
+                app:layout_constraintTop_toBottomOf="@id/etv_mypage_behavior_tag" />
 
             <EditText
                 android:id="@+id/etv_mypage_behavior_review"
