+++ app/src/main/res/layout/activity_mypage_daily.xml
@@ -138,10 +138,40 @@
                 android:layout_height="1dp"
                 android:layout_marginHorizontal="30dp"
                 android:background="@color/black"
-                app:layout_constraintBottom_toTopOf="@id/tv_mypage_daily_title"
                 app:layout_constraintEnd_toEndOf="@id/view_mypage_daily"
                 app:layout_constraintStart_toStartOf="@id/view_mypage_daily"
-                app:layout_constraintTop_toBottomOf="@id/iv_mypage_daily_picture_left" />
+                app:layout_constraintTop_toBottomOf="@id/view_mypage_daily_line_bottom"
+                />
+
+            <TextView
+                android:id="@+id/tv_mypage_daily_tag"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:text="태그(최대 3줄)"
+                android:fontFamily="@font/cafe24"
+                android:textColor="@color/black"
+                android:textSize="14sp"
+                android:layout_marginStart="20dp"
+                android:layout_marginTop="20dp"
+                app:layout_constraintTop_toBottomOf="@id/view_mypage_daily_line_bottom"
+                app:layout_constraintStart_toStartOf="@id/view_mypage_daily"
+                />
+
+            <EditText
+                android:id="@+id/etv_mypage_daily_tag"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:layout_marginHorizontal="50dp"
+                android:layout_marginTop="10dp"
+                android:background="@drawable/mypage_diary_edittext"
+                android:fontFamily="@font/cafe24"
+                android:hint="태그를 작성해주세요"
+                android:padding="20dp"
+                android:textColor="@color/grey"
+                android:textSize="13sp"
+                app:layout_constraintEnd_toEndOf="@id/view_mypage_daily"
+                app:layout_constraintStart_toStartOf="@id/view_mypage_daily"
+                app:layout_constraintTop_toBottomOf="@id/tv_mypage_daily_tag" />
 
             <TextView
                 android:id="@+id/tv_mypage_daily_title"
@@ -154,7 +184,7 @@
                 android:textColor="@color/black"
                 android:textSize="14sp"
                 app:layout_constraintStart_toStartOf="@id/view_mypage_daily"
-                app:layout_constraintTop_toBottomOf="@id/view_mypage_daily_line_bottom" />
+                app:layout_constraintTop_toBottomOf="@id/etv_mypage_daily_tag" />
 
             <EditText
                 android:id="@+id/etv_mypage_daily"
